student manage classes complete + prof del classes debugged + admin courses listview made

+++ src/Model/Student.java
@@ -45,7 +45,7 @@ public class Student implements Serializable {
         return GPA;
     }
 
-    public void setGPA(int GPA) {
+    public void setGPA(float GPA) {
         this.GPA = GPA;
     }
